# resources/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "WuKongIM API",
    "description": "WuKongIM is a high-performance instant messaging system. This API documentation covers all REST endpoints for managing users, channels, messages, connections, and system operations.",
    "version": "2.0.0",
    "contact": {
      "name": "WuKongIM Team",
      "url": "https://github.com/WuKongIM/WuKongIM"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5001",
      "description": "Local development server"
    },
    {
      "url": "https://api.wukongim.com",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/message/send": {
      "post": {
        "summary": "Send Message",
        "description": "Send a message to a channel",
        "tags": ["Message"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessageSendRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Message sent successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageSendResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/message/sync": {
      "post": {
        "summary": "Message Sync (Deprecated)",
        "description": "Message synchronization in write mode (will be deprecated)",
        "deprecated": true,
        "tags": ["Message"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessageSyncRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Messages synchronized successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MessageResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/message/syncack": {
      "post": {
        "summary": "Message Sync Acknowledgment (Deprecated)",
        "description": "Message synchronization acknowledgment in write mode (will be deprecated)",
        "deprecated": true,
        "tags": ["Message"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessageSyncAckRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sync acknowledgment processed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/message/events:append": {
      "post": {
        "summary": "Append Message Event",
        "description": "Service-only append for a typed public message event. The committed anchor is matched by channel, message id, client message number, sender, run id, and authorization fence. `delta` and `snapshot` update the Slot-leader cache; `finish` persists the compact terminal projection through Raft.",
        "tags": ["Message Event"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventAppendRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Event appended successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "data": {
                      "$ref": "#/components/schemas/EventAppendResponse"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request. Possible causes include malformed typed payload, anchor mismatch, stale authority sequence, or a terminal run.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/channel/messagesync": {
      "post": {
        "summary": "Sync Channel Messages",
        "description": "Synchronize messages from a channel with pagination support. Responses can include merged typed event state (`event_meta`).",
        "tags": ["Message"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChannelMessageSyncRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Channel messages synchronized successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChannelMessageSyncResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "ok"
          }
        }
      },
...
      "MessageSendRequest": {
        "type": "object",
        "required": ["payload"],
        "properties": {
          "header": {
            "$ref": "#/components/schemas/MessageHeader"
          },
          "client_msg_no": {
            "type": "string",
            "example": "client_msg_123"
          },
          "stream_no": {
            "type": "string",
            "example": "stream_456"
          },
          "from_uid": {
            "type": "string",
            "example": "user123"
          },
          "channel_id": {
            "type": "string",
            "example": "group123"
          },
          "channel_type": {
            "type": "integer",
            "format": "int8",
            "example": 2
          },
          "expire": {
            "type": "integer",
            "format": "int32",
            "example": 0,
            "description": "Message expiration time in seconds"
          },
          "subscribers": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Specific subscribers to send message to"
          },
          "payload": {
            "type": "string",
            "format": "byte",
            "example": "SGVsbG8gV29ybGQ=",
            "description": "Base64 encoded message content"
          },
          "tag_key": {
            "type": "string",
            "example": "important"
          }
        }
      },
      "MessageSendResponse": {
        "type": "object",
        "properties": {
          "message_id": {
            "type": "integer",
            "format": "int64",
            "example": 123456789
          },
          "message_seq": {
            "type": "integer",
            "format": "int64",
            "example": 1001
          },
          "client_msg_no": {
            "type": "string",
            "example": "client_msg_123"
          }
        }
      },
      "MessageHeader": {
        "type": "object",
        "properties": {
          "no_persist": {
            "type": "integer",
            "enum": [0, 1],
            "example": 0,
            "description": "Whether message should not be persisted (0=persist, 1=no persist)"
          },
          "red_dot": {
            "type": "integer",
            "enum": [0, 1],
            "example": 1,
            "description": "Whether to show red dot notification (0=no, 1=yes)"
          },
          "sync_once": {
            "type": "integer",
            "enum": [0, 1],
            "example": 0,
            "description": "Whether message should be synced only once (0=no, 1=yes)"
          }
        }
      },
      "MessageResponse": {
        "type": "object",
        "properties": {
          "header": {
            "$ref": "#/components/schemas/MessageHeader"
          },
          "setting": {
            "type": "integer",
            "format": "uint8",
            "description": "Message setting flags.",
            "example": 0
          },
          "message_id": {
            "type": "integer",
            "format": "int64",
            "example": 123456789
          },
          "message_idstr": {
            "type": "string",
            "description": "String representation of message_id.",
            "example": "123456789"
          },
          "message_seq": {
            "type": "integer",
            "format": "uint64",
            "example": 1001
          },
          "client_msg_no": {
            "type": "string",
            "example": "client_msg_123"
          },
          "event_meta": {
            "$ref": "#/components/schemas/MessageEventMeta"
          },
          "from_uid": {
            "type": "string",
            "example": "user123"
          },
          "channel_id": {
            "type": "string",
            "example": "group123"
          },
          "channel_type": {
            "type": "integer",
            "format": "uint8",
            "example": 2
          },
          "topic": {
            "type": "string",
            "example": "topic_1"
          },
          "expire": {
            "type": "integer",
            "format": "uint32",
            "example": 0
          },
          "timestamp": {
            "type": "integer",
            "format": "int32",
            "example": 1640995200
          },
          "payload": {
            "type": "string",
            "format": "byte",
            "example": "SGVsbG8gV29ybGQ="
          }
        }
      },
      "MessageEventMeta": {
        "type": "object",
        "properties": {
          "has_events": {
            "type": "boolean",
            "description": "Whether this message has associated event states.",
            "example": true
          },
          "event_version": {
            "type": "integer",
            "format": "uint64",
            "description": "Version of event projection (currently equals last_msg_event_seq).",
            "example": 12
          },
          "last_msg_event_seq": {
            "type": "integer",
            "format": "uint64",
            "description": "Largest msg_event_seq among all event keys.",
            "example": 12
          },
          "event_key_count": {
            "type": "integer",
            "description": "Count of event keys associated with this message.",
            "example": 2
          },
          "open_event_key_count": {
            "type": "integer",
            "description": "Count of event keys still in open state.",
            "example": 1
          },
          "event_keys": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MessageEventKeyMeta"
            }
          }
        }
      },
      "MessageEventKeyMeta": {
        "type": "object",
        "properties": {
          "event_key": {
            "type": "string",
            "description": "Event key identifier inside the message.",
            "example": "main"
          },
          "status": {
            "type": "string",
            "description": "Current status of this event key projection.",
            "enum": ["open", "closed", "error", "cancelled"],
            "example": "open"
          },
          "last_msg_event_seq": {
            "type": "integer",
            "format": "uint64",
            "description": "Last msg_event_seq applied to this event key.",
            "example": 12
          },
          "snapshot": {
            "description": "Optional merged snapshot payload (present in full summary mode).",
            "nullable": true
          },
          "end_reason": {
            "type": "integer",
            "format": "uint8",
            "description": "Terminal reason code for the typed event lane.",
            "example": 0
          },
          "error": {
            "type": "string",
            "description": "Public terminal error code or text when present.",
            "example": ""
          }
        }
      },
...
      "ChannelMessageSyncRequest": {
        "type": "object",
        "required": ["login_uid", "channel_id", "channel_type"],
        "properties": {
          "login_uid": {
            "type": "string",
            "description": "Current logged in user ID",
            "example": "user123"
          },
          "channel_id": {
            "type": "string",
            "description": "Channel ID",
            "example": "channel456"
          },
          "channel_type": {
            "type": "integer",
            "description": "Channel type (1=person, 2=group)",
            "enum": [1, 2]
          },
          "start_message_seq": {
            "type": "integer",
            "description": "Start message sequence (inclusive)",
            "example": 100
          },
          "end_message_seq": {
            "type": "integer",
            "description": "End message sequence (exclusive)",
            "example": 200
          },
          "limit": {
            "type": "integer",
            "description": "Maximum number of messages to return",
            "example": 50,
            "maximum": 10000
          },
          "pull_mode": {
            "type": "integer",
            "description": "Pull mode (0=down, 1=up)",
            "enum": [0, 1]
          },
          "include_event_meta": {
            "type": "integer",
            "description": "Set to 1 to include compact typed event state and snapshots.",
            "example": 1,
            "enum": [0, 1]
          }
        }
      },
      "ChannelMessageSyncResponse": {
        "type": "object",
        "properties": {
          "start_message_seq": {
            "type": "integer",
            "description": "Start message sequence",
            "example": 100
          },
          "end_message_seq": {
            "type": "integer",
            "description": "End message sequence",
            "example": 200
          },
          "more": {
            "type": "integer",
            "description": "Whether there are more messages (1=yes, 0=no)",
            "enum": [0, 1]
          },
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MessageResponse"
            }
          }
        }
      },
...
      "MessageSyncRequest": {
        "type": "object",
        "required": ["uid"],
        "properties": {
          "uid": {
            "type": "string",
            "description": "User ID",
            "example": "user123"
          },
          "message_seq": {
            "type": "integer",
            "format": "uint64",
            "description": "Client's maximum message sequence number for incremental sync",
            "example": 12345
          },
          "limit": {
            "type": "integer",
            "description": "Maximum number of messages to sync",
            "example": 200,
            "default": 200
          }
        }
      },
      "MessageSyncAckRequest": {
        "type": "object",
        "required": ["uid", "last_message_seq"],
        "properties": {
          "uid": {
            "type": "string",
            "description": "User ID",
            "example": "user123"
          },
          "last_message_seq": {
            "type": "integer",
            "description": "Last message sequence number acknowledged",
            "example": 12345
          }
        }
      },
...
      "EventAppendRequest": {
        "type": "object",
        "required": ["channel_id", "channel_type", "from_uid", "message_id", "client_msg_no", "run_id", "authorization_fence", "authority_sequence", "event_id", "event_type", "payload"],
        "description": "Append one typed public event to an already committed AgentRun anchor.",
        "properties": {
          "channel_id": {
            "type": "string",
            "description": "Channel ID. For person-to-person channels, this is the peer's UID.",
            "example": "user_002"
          },
          "channel_type": {
            "type": "integer",
            "description": "Channel type. 1 = person-to-person, 2 = group.",
            "example": 1
          },
          "from_uid": {
            "type": "string",
            "description": "Exact sender UID of the committed anchor.",
            "example": "user_001"
          },
          "message_id": {
            "type": "integer",
            "format": "int64",
            "description": "Exact committed anchor message ID.",
            "example": 123456789
          },
          "client_msg_no": {
            "type": "string",
            "description": "Client message number of the anchor message. The message must already exist.",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "run_id": {
            "type": "string",
            "description": "AgentRun identifier bound into the anchor and every event."
          },
          "authorization_fence": {
            "type": "integer",
            "format": "uint64"
          },
          "authority_sequence": {
            "type": "integer",
            "format": "uint64",
            "minimum": 1
          },
          "event_id": {
            "type": "string",
            "description": "Unique event identifier used for idempotency.",
            "example": "evt_abc123"
          },
          "event_type": {
            "type": "string",
            "description": "Typed event kind.",
            "example": "delta",
            "enum": ["open", "delta", "snapshot", "finish"]
          },
          "event_key": {
            "type": "string",
            "description": "Event key for partitioning multiple event lanes within the same message. Defaults to `main`.",
            "example": "main",
            "default": "main"
          },
          "visibility": {
            "type": "string",
            "description": "Protocol v6 supports public projection only.",
            "example": "public",
            "enum": ["public"]
          },
          "occurred_at": {
            "type": "integer",
            "format": "int64",
            "description": "Event timestamp in milliseconds since epoch. Auto-filled with current time if omitted.",
            "example": 1709827200000
          },
          "payload": {
            "type": "object",
            "description": "Typed AgentRun payload. Delta carries `text_delta`; snapshot and finish carry `snapshot`.",
            "example": {"event_id": "evt_abc123", "run_id": "run_1", "event_key": "main", "event_type": "delta", "authority_sequence": 1, "text_delta": "Hello"}
          },
          "headers": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "description": "Optional key-value headers attached to the event."
          }
        }
      },
      "EventAppendResponse": {
        "type": "object",
        "properties": {
          "client_msg_no": {
            "type": "string",
            "description": "Client message number.",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "event_key": {
            "type": "string",
            "description": "Normalized event key (e.g. empty -> `main`).",
            "example": "main"
          },
          "event_id": {
            "type": "string",
            "description": "Event ID echoed back.",
            "example": "evt_abc123"
          },
          "msg_event_seq": {
            "type": "integer",
            "format": "uint64",
            "description": "Message-level event sequence number. For delta events (cache path), this is the last persisted seq. For persisted events, this is the newly allocated seq.",
            "example": 5
          },
          "event_status": {
            "type": "string",
            "description": "Current event-key status after applying this event.",
            "enum": ["open", "closed", "error", "cancelled"],
            "example": "open"
          },
          "channel_id": {
            "type": "string",
            "description": "Channel ID (original, not fakeChannelID).",
            "example": "user_002"
          },
          "channel_type": {
            "type": "integer",
            "description": "Channel type.",
            "example": 1
          },
          "from_uid": {
            "type": "string",
            "description": "Sender UID.",
            "example": "user_001"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message",
            "example": "Invalid request"
          },
          "message": {
            "type": "string",
            "description": "Detailed error description",
            "example": "The provided parameters are invalid"
          }
        }
      }
    }
  },
...
}